backend/requirements.txt
asgiref==3.10.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.7
djangorestframework==3.16.1
filelock==3.18.0
fsspec==2025.7.0
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
sqlparse==0.5.3
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-decouple>=3.8
requests>=2.31.0